change tim precision to ms + use id instead of ts

+++ cmd/api/handler.go
@@ -23,7 +23,7 @@ type handler struct {
 	account.TokenService
 
 	port      uint
-	tolerance uint64
+	tolerance int64
 }
 
 func (h *handler) Dial(c Config) error {
@@ -75,11 +75,11 @@ func (h *handler) handle(ctx context.Context, raw []byte) error {
 	go h.Send(raw, addr)
 
 	// #Check TS in tolerance range.
-	now := uint64(time.Now().Unix())
+	now := time.Now().Unix()
 	ts := msg.ID.Time()
 	if ts > now || now-ts < h.tolerance {
 		err := gerrors.ErrInvalidTS
-		logger.Error().Err(err).Str("status", "timeout").Uint64("ts", ts).Uint64("now", now).Msg("packet rejected")
+		logger.Error().Err(err).Str("status", "timeout").Int64("ts", ts).Int64("now", now).Msg("packet rejected")
 		return err
 	}
